Removed rootFile:
On line 58 you assign a value to the variable "root", later used as
path prefix to remove when making paths relative. When the command
line option "-d <rootFile>" is given the variable "rootFile" is
assigned the value of the -d option (line 73), but "root" is not
changed!  It is the variable "root" that needs to be updated with the
value of the -d option, "rootFile" is not used except at the very
begining of the main()-method.

I did not notice this when trying out your change before since we are
actually calling the Index.main()-method several times (without
reloading the class) and all calls except the first one with a -d
option will work since "root" will be updated from "rootFile" which
holds the old value of -d from the previous call (allways same value
of -d)...