Close resource. keeping a JarFile open will typically fail renames/deletes under windows
Tracked this from issue with adding file to repositories in Bndtools, via Bnd (R5RepoContentProvider) to Bindex2 that I guess is now RepoIndex? Anyway, keeping a JarFile open will typically fail renames/deletes under windows. It is covered with the existing tests as I was able to simply reproduce it and validate that this simple patch solves it.